+++ irs_reader/dir_utils.py
@@ -1,14 +1,6 @@
-import errno
 import os
 
 
 def mkdir_p(paths):
-    """ Makedirs, from http://stackoverflow.com/a/600612 """
     for path in paths:
-        try:
-            os.makedirs(path)
-        except OSError as exc:  # Python >2.5
-            if exc.errno == errno.EEXIST and os.path.isdir(path):
-                pass
-            else:
-                raise
+        os.makedirs(path, exist_ok=True)