+++ irs_reader/type_utils.py
@@ -1,11 +1,8 @@
 from collections import OrderedDict
 
-# Type names for comparison
-# Checking types is a bad idea, blahdy blahdy blahdy
-
 dictType = type(dict())
 orderedDictType = type(OrderedDict())
 listType = type(list())
-unicodeType = type(u'')
+unicodeType = str
 noneType = type(None)
-strType = type('')
+strType = str